New Year's Resolutions
Another year comes to a close and with that an opportunity for new beginings - and New Year's Resolutions is an opportunity to do just that.

At the same time, in a 2014 report, 35% of participants who failed their New Year's Resolutions admitted they had unrealistic goals, 33% of participants did not keep track of their progress, and 23% forgot about them; about one in 10 respondents claimed they made too many resolutions. [1]

A 2007 study from the University of Bristol involving 3,000 people showed that 88% of those who set New Year resolutions fail, despite the fact that 52% of the study's participants were confident of success at the beginning. [2]

With this dataset, containing 5011 tweets of new year's resolutions, you can use the collective knowledge to improve your odds of success in your own resolutions!